requirements.txt
aiogram==3.1.0
uvicorn==0.23.2
starlette==0.30.1
python-dotenv==1.0.1